Renamed file to match Kirktopode's refactoring
